+++ packages/orq-rc/src/models/components/evaluatorresponsejsonschema.ts
@@ -28,8 +28,8 @@ export const EvaluatorResponseJsonSchema$inboundSchema: z.ZodType<
 > = z.object({
   _id: z.string(),
   description: z.string(),
-  created: z.string().default("2026-07-21T15:01:26.715Z"),
-  updated: z.string().default("2026-07-21T15:01:26.715Z"),
+  created: z.string().default("2026-07-21T15:24:50.532Z"),
+  updated: z.string().default("2026-07-21T15:24:50.532Z"),
   updated_by_id: z.nullable(z.string()).optional(),
   guardrail_config: z.any().optional(),
   type: z.literal("json_schema"),