+++ packages/orq-rc/RELEASES.md
@@ -11019,3 +11019,13 @@ Based on:
 - [typescript v4.12.0-rc.48] packages/orq-rc
 ### Releases
 - [NPM v4.12.0-rc.48] https://www.npmjs.com/package/@orq-ai/node/v/4.12.0-rc.48 - packages/orq-rc
+
+## 2026-07-21 15:33:02
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.790.3 (2.918.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.12.0-rc.49] packages/orq-rc
+### Releases
+- [NPM v4.12.0-rc.49] https://www.npmjs.com/package/@orq-ai/node/v/4.12.0-rc.49 - packages/orq-rc